backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
python-multipart==0.0.18
anthropic==0.40.0
openai==1.57.0
pdfminer.six==20231228
python-docx==1.1.2
httpx==0.28.0
python-jose==3.3.0
bcrypt==4.2.1
aiofiles==24.1.0
